TSK-1143 even more comments from Holger
